1- Moved serialization utils to a sub-package 2- Fixed SshRunner 3- Added extra option to the nvidia-smi demo shim
