+++ py/genv/serialization/partial_deserialization.py
