+++ py/genv/snapshot.py
@@ -6,9 +6,9 @@ from datetime import datetime
 from . import processes as processes_
 from . import envs as envs_
 from . import devices as devices_
-from .partial_deserialization import smart_ctor
-from .runners.local import LocalRunner
-from .runners.runner import Runner
+from genv.serialization.partial_deserialization import smart_ctor
+from .runners import LocalRunner
+from .runners import Runner
 
 
 @dataclass
